extra logica voor api

+++ web/projectantwerpen/app/Http/Controllers/IndividualProjectController.php
@@ -9,6 +9,7 @@ use App\Comment;
 use App\Http\Controllers\Controller;
 use Auth;
 use Redirect;
+use Session;
 
 class IndividualProjectController extends Controller
 {
@@ -29,26 +30,29 @@ class IndividualProjectController extends Controller
 
 	public function like($id) {
 		$project = Project::find($id);
-		$user_id = Auth::user()->id;
-		if(Auth::user()){
-			$like = Users_project::where('fk_user', $user_id)->where('fk_project', $id)->where('like', true)->first();
-			$dislike = Users_project::where('fk_user', $user_id)->where('fk_project', $id)->where('like', false)->first();
-			var_dump(('like'));
-			if($dislike){
-
-				$dislike->delete();
-				$users_project = Users_project::create(['fk_user' => $user_id, 'fk_project' => $id, 'like' => true]);
-				return Redirect::back();
+		if(Session::get("id") != $id){
+			if(Auth::user()){
+				$user_id = Auth::user()->id;
+				$like = Users_project::where('fk_user', $user_id)->where('fk_project', $id)->where('like', true)->first();
+				$dislike = Users_project::where('fk_user', $user_id)->where('fk_project', $id)->where('like', false)->first();
+				if($dislike){
+					$dislike->delete();
+					$users_project = Users_project::create(['fk_user' => $user_id, 'fk_project' => $id, 'like' => true]);
+					return Redirect::back();
+				}
+				elseif($like){
+					return Redirect::back();
+				}
+				elseif(!$like && !$dislike){
+					$users_project = Users_project::create(['fk_user' => $user_id, 'fk_project' => $id, 'like' => true]);
+					return Redirect::back();
+				}
 			}
-			elseif($like){
-				return Redirect::back();
-			}
-			elseif(!$like && !$dislike){
-				$users_project = Users_project::create(['fk_user' => $user_id, 'fk_project' => $id, 'like' => true]);
-				return Redirect::back();
+			else{
+				$anon_like = Users_project::create(['fk_user' => null, 'fk_project' => $id, 'like' => true]);
+				Session::put(["id" => $id]);
 			}
 		}
-		
 		$likes = $project->likes($id);
 		$dislikes = $project->dislikes($id);
 		$comments = $project->comments($id)->with('user')->get();
@@ -58,24 +62,30 @@ class IndividualProjectController extends Controller
 
 	public function dislike($id) {
 		$project = Project::find($id);
-		$user_id = Auth::user()->id;
-		if(Auth::user()){
-			$like = Users_project::where('fk_user', $user_id)->where('fk_project', $id)->where('like', true)->first();
-			$dislike = Users_project::where('fk_user', $user_id)->where('fk_project', $id)->where('like', false)->first();
-			if($like){
-				$like->delete();
-				$users_project = Users_project::create(['fk_user' => $user_id, 'fk_project' => $id, 'like' => false]);
-				return Redirect::back();
-			}
-			elseif($dislike){
-				return Redirect::back();
+		if(Session::get("id") != $id){
+			if(Auth::user()){
+				$user_id = Auth::user()->id;
+				$like = Users_project::where('fk_user', $user_id)->where('fk_project', $id)->where('like', true)->first();
+				$dislike = Users_project::where('fk_user', $user_id)->where('fk_project', $id)->where('like', false)->first();
+				if($like){
+					$like->delete();
+					$users_project = Users_project::create(['fk_user' => $user_id, 'fk_project' => $id, 'like' => false]);
+					return Redirect::back();
+				}
+				elseif($dislike){
+					return Redirect::back();
+				}
+				else{
+					$users_project = Users_project::create(['fk_user' => $user_id, 'fk_project' => $id, 'like' => false]);
+					return Redirect::back();
+				}
 			}
 			else{
-				$users_project = Users_project::create(['fk_user' => $user_id, 'fk_project' => $id, 'like' => false]);
-				return Redirect::back();
+				$anon_dislike = Users_project::create(['fk_user' => null, 'fk_project' => $id, 'like' => false]);
+			
+				Session::put(["id" => $id]);
 			}
 		}
-		
 		$likes = $project->likes($id);
 		$dislikes = $project->dislikes($id);
 		$comments = $project->comments($id)->with('user')->get();
